+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0]
+
+* change verify_ssl config to MailChimp class (from MailChimpSignupPage)
+* add campaign listing page
+
 ## [2.4.1]
 
 * fix typo in verify_ssl config